+++ tests/utils.py
@@ -143,6 +143,7 @@ STATUS_CODE_TO_ROUTE_DEFAULT = [
             "/profile/<username>",
             "/auth/logout",
             "/auth/reset_password/<token>",
+            "/favicon.ico",
         ],
     ),
     (